mcp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.21.0
google-auth==2.35.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
google-adk
google-generativeai==0.8.3
jinja2==3.1.4
weasyprint==64.0
matplotlib
google-cloud-storage==2.17.0